requirements.txt
streamlit
requests
beautifulsoup4
textstat
sentence-transformers
langchain
faiss-cpu
plotly
transformers
torch